chore: improve test coverage for umi-build-dev
