+++ packages/umi-build-dev/src/getConfig/test/base.test.js
@@ -1,0 +1,8 @@
+import base from '../configPlugins/base';
+
+test('validate', () => {
+  base().validate('abc');
+  expect(() => {
+    base().validate(1);
+  }).toThrow(/Configure item base should be String/);
+});
