+++ packages/umi-build-dev/src/importsToStr.js
@@ -4,7 +4,7 @@ export default function(imports) {
   return imports.map(imp => {
     const { source, specifier } = imp;
     if (specifier) {
-      return `import ${specifier} from '${winPath(source)}'`;
+      return `import ${specifier} from '${winPath(source)}';`;
     } else {
       return `import '${winPath(source)}';`;
     }
