+++ packages/umi-build-dev/src/importsToStr.test.js
@@ -1,0 +1,8 @@
+import importsToStr from './importsToStr';
+
+test('importsToStr', () => {
+  expect(importsToStr([{ source: 'foo' }, { source: 'bar', specifier: 'bar' }])).toEqual([
+    `import 'foo';`,
+    `import bar from 'bar';`,
+  ]);
+});
